+++ src/syntax.rs
@@ -1,7 +1,6 @@
 pub mod ast;
 pub mod op;
 pub mod tokenizer;
-pub mod types;
 
 use crate::{
     src_file::{Position, SrcFile},